refactor: move GROUP_PERMISSION_NAMES to constants file
