+++ src/components/Proposals/ProposalActions/ProposeRoleModal.tsx
@@ -31,9 +31,11 @@ import {
   ProposalActionType,
   ProposeRoleModalFieldName,
 } from "../../../constants/proposal.constants";
-import { DEFAULT_ROLE_COLOR } from "../../../constants/role.constants";
+import {
+  DEFAULT_ROLE_COLOR,
+  GROUP_PERMISSION_NAMES,
+} from "../../../constants/role.constants";
 import { initGroupRolePermissions } from "../../../utils/role.utils";
-import { GROUP_PERMISSION_NAMES } from "../../Groups/GroupRoles/GroupPermissionsForm";
 import Accordion, {
   AccordionDetails,
   AccordionSummary,
